backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
torch==2.0.1
torchvision==0.15.2
spacy==3.7.2
transformers==4.30.2
opencv-python==4.8.1.78
numpy==1.24.4
Pillow==10.0.1
requests==2.31.0
nltk==3.9.1